Refactor DocBlockResolver creates DocBlockResolvable dto
